Coloured sky and ground... but not quite right

+++ pfd-spkfun-tft/point.h
@@ -1,5 +1,5 @@
 // The Ardiuno IDE is dumb: typedefs MUST be in header files.
 
 typedef struct {
-    int16_t x, y;
+    int x, y;
 } point;
